+++ packages/adaptive_style/lib/adaptive_style.dart
@@ -1,0 +1,15 @@
+library adaptive_style;
+
+export 'src/adaptive_achor_positioned.dart';
+export 'src/scale_mediaquery_widget.dart';
+export 'src/adaptive_stack.dart';
+export 'src/adaptive_widget_builder.dart';
+export 'src/adaptive_widget.dart';
+export 'src/device_size.dart';
+export 'src/extension.dart';
+export 'src/inherited_scale_ref.dart';
+export 'src/scale_ref_provider.dart';
+export 'src/scale_ref.dart';
+export 'src/size_scale.dart';
+export 'src/anchor_edges.dart';
+export 'src/positioned_data.dart';