+++ packages/adaptive_style/lib/src/scale_ref.dart
@@ -1,0 +1,6 @@
+import 'size_scale.dart';
+import 'device_size.dart';
+
+/// a record representing the closest supported size
+/// and the scale between the real size and the supported size
+typedef ScaleRef = ({DeviceSize idealSize, SizeScale scale});