+++ packages/yak_flutter/lib/src/widgets/all.dart
@@ -1,2 +1,7 @@
-export 'preferred_size_themed.dart';
 export 'edge_positioned.dart';
+export 'inherited_value_notifier.dart';
+export 'media_query_widget.dart';
+export 'preferred_size_themed.dart';
+export 'scalable_edge_positioned.dart';
+export 'scalable_sized_positioned.dart';
+export 'sized_positioned.dart';